requirements.txt
numpy
opencv-python
pynetworktables
robotpy-cscore
